requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.27.0
openai==1.54.3
jiter==0.12.0
numpy==1.26.4
pgvector==0.3.6
neo4j==5.15.0
structlog==24.1.0
pytest==8.0.0
pytest-asyncio==0.23.0
